+++ Changelog.md
@@ -14,8 +14,14 @@
 
 ## Unreleased
 
+MSRV bumped to 1.79.
+
+Now `quick-xml` supports the UTF-16 encoded documents. See the new `DecodingReader` type.
+
 ### New Features
 
+- [#956]: Add `DecodingReader`, a `BufRead` adapter that auto-detects encoding
+  from BOM or XML declaration and transcodes to UTF-8. Enabled by the `encoding` feature.
 - [#938]: Add new enumeration `XmlVersion` and typified getter `BytesDecl::xml_version()`.
 - [#938]: Add new error variant `IllFormedError::UnknownVersion`.
 - [#371]: Add new error variant `EscapeError::TooManyNestedEntities`.
@@ -46,11 +52,13 @@
   accepts `XmlVersion` parameter to apply correct EOL normalization rules.
 - [#944]: `read_text()` now returns `BytesText` which allows you to get the content with
   properly normalized EOLs. To get the previous behavior use `.read_text().decode()?`.
+- [#956]: Bumped MSRV from 1.59 (Feb 2022) to 1.79 (June 2024)
 
 [#371]: https://github.com/tafia/quick-xml/issues/371
 [#914]: https://github.com/tafia/quick-xml/pull/914
 [#938]: https://github.com/tafia/quick-xml/pull/938
 [#944]: https://github.com/tafia/quick-xml/pull/944
+[#956]: https://github.com/tafia/quick-xml/pull/956
 
 
 ## 0.39.4 -- 2026-05-08